tax invoice fixes, view order from enquiry details screen fix

+++ app/src/main/java/com/adrosonic/craftexchange/database/entities/realmEntities/TaxInvDetails.kt
@@ -18,8 +18,8 @@ open class TaxInvDetails : RealmObject() {
     var invoiceNo: String?=""
     var advancePaidAmt: Long?=0
     var deliveryCharges: Long?=0
-    var cgst: Long?=0
-    var sgst: Long?=0
+    var cgst: Double?=0.0
+    var sgst: Double?=0.0
     var finalTotalAmt: Double?=0.0
     var isactive: Long?=null
     var createdon: String?=""
